+++ packages/block-library/src/post-comment/edit.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { __ } from '@wordpress/i18n';
+import { __, _x } from '@wordpress/i18n';
 import { Placeholder, TextControl, Button } from '@wordpress/components';
 import { useState } from '@wordpress/element';
 import { blockDefault } from '@wordpress/icons';
@@ -20,8 +20,7 @@ const TEMPLATE = [
 	[ 'core/post-comment-author' ],
 ];
 
-export default function Edit( { attributes, setAttributes } ) {
-	const { commentId } = attributes;
+export default function Edit( { attributes: { commentId }, setAttributes } ) {
 	const [ commentIdInput, setCommentIdInput ] = useState( commentId );
 	const blockProps = useBlockProps();
 	const innerBlocksProps = useInnerBlocksProps( blockProps, {
@@ -34,7 +33,7 @@ export default function Edit( { attributes, setAttributes } ) {
 			<div { ...blockProps }>
 				<Placeholder
 					icon={ blockDefault }
-					label={ __( 'Post Comment' ) }
+					label={ _x( 'Post Comment', 'block title' ) }
 					instructions={ __(
 						'To show a comment, input the comment ID.'
 					) }