+++ packages/components/src/color-indicator/stories/index.js
@@ -11,6 +11,9 @@ import ColorIndicator from '../';
 export default {
 	title: 'Components/ColorIndicator',
 	component: ColorIndicator,
+	parameters: {
+		knobs: { disabled: false },
+	},
 };
 
 export const _default = () => {