+++ packages/components/src/color-picker/test/index.js
@@ -41,18 +41,6 @@ function moveReactColorfulSlider( sliderElement, from, to ) {
 	fireEvent( sliderElement, new FakeMouseEvent( 'mousemove', to ) );
 }
 
-const sleep = ( ms ) => {
-	const promise = new Promise( ( resolve ) => setTimeout( resolve, ms ) );
-	jest.advanceTimersByTime( ms + 1 );
-	return promise;
-};
-
-const hslMatcher = expect.objectContaining( {
-	h: expect.any( Number ),
-	s: expect.any( Number ),
-	l: expect.any( Number ),
-} );
-
 const hslaMatcher = expect.objectContaining( {
 	h: expect.any( Number ),
 	s: expect.any( Number ),
@@ -61,7 +49,6 @@ const hslaMatcher = expect.objectContaining( {
 } );
 
 const legacyColorMatcher = {
-	color: expect.anything(),
 	hex: expect.any( String ),
 	hsl: hslaMatcher,
 	hsv: expect.objectContaining( {
@@ -100,23 +87,15 @@ describe( 'ColorPicker', () => {
 				{ pageX: 10, pageY: 10 }
 			);
 
-			// `onChange` is debounced so we need to sleep for at least 1ms before checking that onChange was called
-			await sleep( 1 );
-
 			expect( onChangeComplete ).toHaveBeenCalledWith(
 				legacyColorMatcher
 			);
 		} );
 	} );
 
-	it( 'should fire onChange with the HSLA value', async () => {
+	it( 'should fire onChange with the string value', async () => {
 		const onChange = jest.fn();
-		const color = {
-			h: 125,
-			s: 0.2,
-			l: 0.5,
-			a: 0.5,
-		};
+		const color = 'rgba(1, 1, 1, 0.5)';
 
 		const { container } = render(
 			<ColorPicker onChange={ onChange } color={ color } enableAlpha />
@@ -129,10 +108,9 @@ describe( 'ColorPicker', () => {
 			{ pageX: 10, pageY: 10 }
 		);
 
-		// `onChange` is debounced so we need to sleep for at least 1ms before checking that onChange was called
-		await sleep( 1 );
-
-		expect( onChange ).toHaveBeenCalledWith( hslaMatcher );
+		expect( onChange ).toHaveBeenCalledWith(
+			expect.stringMatching( /^#([a-fA-F0-9]{8})$/ )
+		);
 	} );
 
 	it( 'should fire onChange with the HSL value', async () => {
@@ -160,9 +138,8 @@ describe( 'ColorPicker', () => {
 			{ pageX: 10, pageY: 10 }
 		);
 
-		// `onChange` is debounced so we need to sleep for at least 1ms before checking that onChange was called
-		await sleep( 1 );
-
-		expect( onChange ).toHaveBeenCalledWith( hslMatcher );
+		expect( onChange ).toHaveBeenCalledWith(
+			expect.stringMatching( /^#([a-fA-F0-9]{6})$/ )
+		);
 	} );
 } );