+++ packages/edit-site/src/components/block-editor/index.js
@@ -13,16 +13,12 @@ import {
 	BlockEditorProvider,
 	__experimentalLinkControl,
 	BlockInspector,
-	BlockList,
 	BlockTools,
 	__unstableBlockSettingsMenuFirstItem,
-	__experimentalUseResizeCanvas as useResizeCanvas,
 	__unstableUseTypingObserver as useTypingObserver,
-	__unstableUseMouseMoveTypingReset as useMouseMoveTypingReset,
-	__unstableEditorStyles as EditorStyles,
-	__unstableIframe as Iframe,
+	store as blockEditorStore,
 } from '@wordpress/block-editor';
-import { useMergeRefs } from '@wordpress/compose';
+import { useMergeRefs, useViewportMatch } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -34,28 +30,19 @@ import { store as editSiteStore } from '../../store';
 import BlockInspectorButton from './block-inspector-button';
 import EditTemplatePartMenuButton from '../edit-template-part-menu-button';
 import BackButton from './back-button';
-
-const LAYOUT = {
-	type: 'default',
-	// At the root level of the site editor, no alignments should be allowed.
-	alignments: [],
-};
+import ResizableEditor from './resizable-editor';
 
 export default function BlockEditor( { setIsInserterOpen } ) {
-	const { settings, templateType, page, deviceType } = useSelect(
+	const { settings, templateType, page } = useSelect(
 		( select ) => {
-			const {
-				getSettings,
-				getEditedPostType,
-				getPage,
-				__experimentalGetPreviewDeviceType,
-			} = select( editSiteStore );
+			const { getSettings, getEditedPostType, getPage } = select(
+				editSiteStore
+			);
 
 			return {
 				settings: getSettings( setIsInserterOpen ),
 				templateType: getEditedPostType(),
 				page: getPage(),
-				deviceType: __experimentalGetPreviewDeviceType(),
 			};
 		},
 		[ setIsInserterOpen ]
@@ -65,10 +52,10 @@ export default function BlockEditor( { setIsInserterOpen } ) {
 		templateType
 	);
 	const { setPage } = useDispatch( editSiteStore );
-	const resizedCanvasStyles = useResizeCanvas( deviceType, true );
-	const ref = useMouseMoveTypingReset();
 	const contentRef = useRef();
 	const mergedRefs = useMergeRefs( [ contentRef, useTypingObserver() ] );
+	const isMobileViewport = useViewportMatch( 'small', '<' );
+	const { clearSelectedBlock } = useDispatch( blockEditorStore );
 
 	const isTemplatePart = templateType === 'wp_template_part';
 
@@ -102,21 +89,25 @@ export default function BlockEditor( { setIsInserterOpen } ) {
 					'is-focus-mode': isTemplatePart,
 				} ) }
 				__unstableContentRef={ contentRef }
+				onClick={ ( event ) => {
+					// Clear selected block when clicking on the gray background.
+					if ( event.target === event.currentTarget ) {
+						clearSelectedBlock();
+					}
+				} }
 			>
 				<BackButton />
-				<Iframe
-					style={ resizedCanvasStyles }
-					head={ <EditorStyles styles={ settings.styles } /> }
-					ref={ ref }
+
+				<ResizableEditor
+					enableResizing={
+						isTemplatePart &&
+						// Disable resizing in mobile viewport.
+						! isMobileViewport
+					}
+					settings={ settings }
 					contentRef={ mergedRefs }
-					name="editor-canvas"
-					className="edit-site-visual-editor__editor-canvas"
-				>
-					<BlockList
-						className="edit-site-block-editor__block-list wp-site-blocks"
-						__experimentalLayout={ LAYOUT }
-					/>
-				</Iframe>
+				/>
+
 				<__unstableBlockSettingsMenuFirstItem>
 					{ ( { onClose } ) => (
 						<BlockInspectorButton onClick={ onClose } />